Crate optee_teec

Re-exports

pub use optee_teec_sys as raw;

Structs

Context

An abstraction of the logical connection between a client application and a TEE.

Error

The error type for TEE operations of Context and Session.

Operation

This type defines the payload of either an open session operation or an invoke command operation. It is also used for cancellation of operations, which may be desirable even if no payload is passed.

ParamNone

Represents none parameter which carries no information.

ParamTmpRef

This type defines a temporary memory reference. It is used as a Operation parameter when the corresponding parameter type is one of MemrefTempInput, MemrefTempOutput, or MemrefTempInout.

ParamTypes

ParamValue

This type defines a parameter that is not referencing shared memory, but carries instead small raw data passed by value. It is used as a Operation parameter when the corresponding parameter type is one of ValueInput, ValueOutput, or ValueInout.

PluginMethod

PluginParameters

struct PluginParameters { @cmd: u32, plugin cmd, defined in proto/ @sub_cmd: u32, plugin subcmd, defined in proto/ @inout: &’a mut u8, input/output buffer shared with TA and plugin @outlen, length of output sent to TA }

Session

Represents a connection between a client application and a trusted application.

Uuid

A Universally Unique Resource Identifier (UUID) type as defined in RFC4122. The value is used to identify a trusted application.

Enums

ConnectionMethods

Session login methods.

ErrorKind

A list specifying general categories of TEE client error and its corresponding code in OP-TEE client library.

ErrorOrigin

ParamType

These are used to indicate the type of Parameter encoded inside the operation structure.

Traits

Param

Type Aliases

Result

A specialized Result type for TEE operations.

Attribute Macros

plugin_init

Attribute to declare the init function of a plugin

plugin_invoke

Attribute to declare the invoke function of a plugin